Integrated Clerk for Kanban App, Refactored APIs, and Schema for Individual Data Protection
Description:
In this commit, I integrated Clerk, a secure authentication and user management tool, into my Kanban app. With Clerk, each user now has access to their personalized data, ensuring data privacy and protection. I refactored the APIs and adjusted the schema to enable individual data segregation and to restrict access to unauthorized users.

Key Changes:
Integrated Clerk for secure authentication and user management.
Refactored APIs to handle user-specific data retrieval.
Modified the schema to support individual data protection.
Implemented access controls to safeguard main pages based on user authentication status.

+++ src/server/controllers/tasks.ts
@@ -1,10 +1,11 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 import * as yup from 'yup';
-import { PayloadError } from '@/src/server/libs/customErrors';
+import { HTTPNotAuthorizedError, PayloadError } from '@/src/server/libs/customErrors';
 import prisma from '@/src/libs/prisma';
 import { Prisma } from '@prisma/client';
 import { moveTaskMiddleware } from '@/src/server/middlewares/tasks';
 import { InferType } from 'yup';
+import isAuthenticated from '@/src/server/middlewares/auth';
 
 type Subtask = {
   id: string;
@@ -15,9 +16,12 @@ type Subtask = {
 export async function getTask(req: NextApiRequest, res: NextApiResponse) {
   try {
     const taskId = req.query.taskId as string;
+    const userId = isAuthenticated(req);
+
     const task = await prisma.task.findUniqueOrThrow({
       where: {
-        id: taskId
+        id: taskId,
+        userId
       },
       include: {
         subtasks: true
@@ -26,6 +30,7 @@ export async function getTask(req: NextApiRequest, res: NextApiResponse) {
 
     res.json(task);
   } catch (error: any) {
+    if (error instanceof HTTPNotAuthorizedError) return res.status(401).json(error.getError());
     if (error instanceof Prisma.PrismaClientKnownRequestError)
       return res
         .status(error.code === 'P2025' ? 404 : 400)
@@ -51,8 +56,10 @@ export async function createTask(req: NextApiRequest, res: NextApiResponse) {
     const columnId = req.query.columnId as string;
     const boardId = req.query.boardId as string;
 
+    const userId = isAuthenticated(req);
+
     const column = await prisma.column.findUniqueOrThrow({
-      where: { id: columnId, boardId },
+      where: { id: columnId, boardId, userId },
       select: {
         name: true
       }
@@ -66,16 +73,17 @@ export async function createTask(req: NextApiRequest, res: NextApiResponse) {
       data: {
         ...data,
         subtasks: undefined,
-        columnId
+        columnId,
+        userId
       }
     });
 
     await prisma.subTask.createMany({
-      data: data.subtasks?.map(subtask => ({ taskId: newTask.id, ...subtask })) ?? []
+      data: data.subtasks?.map(subtask => ({ taskId: newTask.id, ...subtask, userId })) ?? []
     });
 
     const task = await prisma.task.findUnique({
-      where: { id: newTask.id },
+      where: { id: newTask.id, userId },
       include: {
         subtasks: true
       }
@@ -84,6 +92,7 @@ export async function createTask(req: NextApiRequest, res: NextApiResponse) {
     res.status(201).json({ message: task!.title + ' task was created.', task });
   } catch (error: any) {
     console.log(error);
+    if (error instanceof HTTPNotAuthorizedError) return res.status(401).json(error.getError());
     if (error instanceof yup.ValidationError) return res.status(400).json(PayloadError.getError(error));
     if (error instanceof Prisma.PrismaClientKnownRequestError)
       return res.status(error.code === 'P2025' ? 404 : 400).json(
@@ -115,6 +124,8 @@ export async function updateTask(req: NextApiRequest, res: NextApiResponse) {
 
     schema.validateSync(data);
 
+    const userId = isAuthenticated(req);
+
     const subtasksWithId: Subtask[] = data.subtasks?.filter((subtask: Subtask) => !!subtask.id) ?? [];
     const subtasksWithoutId: Subtask[] = data.subtasks?.filter((subtask: Subtask) => !subtask.id) ?? [];
 
@@ -124,7 +135,8 @@ export async function updateTask(req: NextApiRequest, res: NextApiResponse) {
           title: subtask.title
         },
         where: {
-          id: subtask.id
+          id: subtask.id,
+          userId
         }
       });
     }
@@ -132,6 +144,7 @@ export async function updateTask(req: NextApiRequest, res: NextApiResponse) {
     await prisma.subTask.deleteMany({
       where: {
         taskId,
+        userId,
         id: {
           notIn: subtasksWithId.map(subtask => subtask.id)
         }
@@ -139,7 +152,7 @@ export async function updateTask(req: NextApiRequest, res: NextApiResponse) {
     });
 
     await prisma.subTask.createMany({
-      data: subtasksWithoutId.map(subtask => ({ taskId, ...subtask }))
+      data: subtasksWithoutId.map(subtask => ({ taskId, ...subtask, userId }))
     });
 
     await moveTaskMiddleware(req);
@@ -149,7 +162,7 @@ export async function updateTask(req: NextApiRequest, res: NextApiResponse) {
         title: data.title,
         description: data.description
       },
-      where: { id: taskId },
+      where: { id: taskId, userId },
       include: {
         subtasks: true
       }
@@ -157,6 +170,7 @@ export async function updateTask(req: NextApiRequest, res: NextApiResponse) {
 
     res.json(task);
   } catch (error: any) {
+    if (error instanceof HTTPNotAuthorizedError) return res.status(401).json(error.getError());
     if (error instanceof Prisma.PrismaClientKnownRequestError)
       return res
         .status(error.code === 'P2025' ? 404 : 400)
@@ -169,6 +183,7 @@ export async function moveTask(req: NextApiRequest, res: NextApiResponse) {
   try {
     res.json(await moveTaskMiddleware(req));
   } catch (error: any) {
+    if (error instanceof HTTPNotAuthorizedError) return res.status(401).json(error.getError());
     if (error instanceof Prisma.PrismaClientKnownRequestError)
       return res.status(error.code === 'P2025' ? 404 : 400).json({
         message: ((error.meta?.cause ?? 'The column you want to move the task to was not found.') as string).replace(
@@ -183,9 +198,12 @@ export async function moveTask(req: NextApiRequest, res: NextApiResponse) {
 export async function deleteTask(req: NextApiRequest, res: NextApiResponse) {
   try {
     const taskId = req.query.taskId as string;
-    const task = await prisma.task.delete({ where: { id: taskId } });
+    const userId = isAuthenticated(req);
+
+    const task = await prisma.task.delete({ where: { id: taskId, userId } });
     res.json({ message: task.title + ' task was deleted.' });
   } catch (error: any) {
+    if (error instanceof HTTPNotAuthorizedError) return res.status(401).json(error.getError());
     if (error instanceof Prisma.PrismaClientKnownRequestError)
       return res
         .status(error.code === 'P2025' ? 404 : 400)
